+++ proxypay/management/commands/proxypay.py
@@ -3,11 +3,12 @@
 #
 
 # django stuff
-from django.core.management.base import BaseCommand, CommandError
+from django.core.management.base import BaseCommand
 
 # proxypay stuff
 from proxypay.api import api
 from proxypay.models import Reference
+from proxypay.conf import PP_CONFIG_DEFAULT_DEVELOPMENT_ENV
 
 # =====================================================================================================================
 
@@ -25,6 +26,10 @@ class Command(BaseCommand):
 
         # test payment
         if args[0] == 'pay':
+
+            if api.env != PP_CONFIG_DEFAULT_DEVELOPMENT_ENV:
+                raise Exception('It is only possible to make fictitious payments in a sandbox environment')
+
             # getting the reference id from args
             reference_id = int(args[1])
             # get Reference model instance
@@ -44,6 +49,8 @@ class Command(BaseCommand):
                     reference.paid(payment)
                     # 
                     self.stdout.write(self.style.SUCCESS(f"Reference: <{reference.reference}>, paid successfully"))
+                    # acknowledge payment
+                    api.acknowledge_payment(payment.get('id'))
                 else:
                     self.stdout.write(self.style.ERROR(f"Error recognizing payment"))
             else: